Add chemical spraying methods.
spraying a chemical to two different shaped fields, say one is triangular and the other is rectangular shaped. You can choose the length of each edge of the fields and encode them into your program. Your program should drive the trejectory of the drone completely covering the field. You may assume that the drove will start flying in a point inside the field.